+++ include/roq/client/event_log_multiplexer.hpp
@@ -15,6 +15,7 @@ struct ROQ_PUBLIC EventLogMultiplexer {
     // stream
     virtual void operator()(Event<StreamStatus> const &) = 0;
     virtual void operator()(Event<ExternalLatency> const &) = 0;
+    virtual void operator()(Event<RateLimitsUpdate> const &) = 0;
     virtual void operator()(Event<RateLimitTrigger> const &) = 0;
 
     // service